fixing format_tools
