+++ module/core/inspect_type/src/lib.rs
@@ -2,7 +2,9 @@
 #![ doc( html_favicon_url = "https://raw.githubusercontent.com/Wandalen/wTools/alpha/asset/img/logo_v3_trans_square_icon_small_v2.ico" ) ]
 #![ doc( html_root_url = "https://docs.rs/inspect_type/latest/inspect_type/" ) ]
 #![ doc = include_str!( concat!( env!( "CARGO_MANIFEST_DIR" ), "/", "Readme.md" ) ) ]
+#![ allow( unexpected_cfgs ) ]
 
+// #[ allow( unexpected_cfgs ) ]
 // #[ cfg( RUSTC_IS_NIGHTLY ) ]
 #[ cfg( not( RUSTC_IS_STABLE ) ) ]
 mod nightly
